[ML] add deployment_stats to trained model stats
